Move AudioManager into audiomanager folder
